Add hacky types.Type
We want to explore encoding type information about Noms data in
the Noms database. So, we need some way to describe types. This
takes the shortest path to making a Noms type called "Type" that
is a peer of Set, Map et al and can describe all the types we currently
use -- except types.Value, which we want to replace in the type system
with Unions (a TODO).