Go: Make ValidatingBatchSink read Values during validation
In order to ensure that pulling works without requiring extensive
Hint gymnastics on the client, we're willing to relax our server-side
requirement that every writeValue request contain all information
needed to complete validation. Now, in the event that the validation
code encounters a Ref that's not in its validation cache, it'll just
go ReadValue() it and soldier on.

Toward #1568