When decoding a Chunk, re-use the hash in the decoded Value
This should mean that we calculate the hash of read Values only one,
when pulling the chunk out of storage. It means we need to be careful
about making sure that we actually do calculate hashes of chunks we
read, or that we get the hash from a trusted place.