Fix 2 deadlocks in orderedSequenceDiffBest causing noms log to hang
First: diff wasn't checking whether it had stopped before sending the
final set of changes. If the caller had stopped - and therefore no
longer reading from the changes channel - diff would hang.

Second: the same close channel was being used on 2 threads and it was
possible for both to read the close signal - causing the other to hang.