Make Pull step down one level at a time, de-dup source DB work queue
During a Pull(), the code can encounter the same Ref a number of times.
To mitigate this most expeditiously, the code now is less aggressive about
parallelizing work. It steps down the tree one ref-height level at a time,
and de-dups across the set of refs it discovers by traversing that level.

Also, it turns out Pull() was making some bad assumptions about how the Go
heap package used its backing storage. Since it wasn't really relying
on heap guarantees anyway, this changes the code to use a slice of Ref
that's sorted in order of increasing ref-height: RefByHeight.

Fixes #2142