JS: Fix encoding bugs related to signed varints
The library that we were using multiplied (and divided) by 2 to
do the zigzag encoding. However, if we are already close to the
precision limit then we lose precision and the encoding/decoding
does not round trip.

Instead, we split the float64 into two Uint32 numbers and do the
operations on those.

Fixes #2104
Fixes #2234