Remove return value from datas.Pull()
Raf realized that we could detect a number of non-fast-forward cases
from the relative ref-height of sink Head and source Head. To cover
the rest of the cases, it's easiest to just attempt a fast forward
and, if it fails with ErrMergeNeeded, go ahead and try setting the
head directly.

Fixes #2240