Let Splore open a noms path
Currently Splore can only open a database (db=) or a hash (hash=). This
removes the hash and replaces it with the much more general path (p=).

For example, what used to be:
  ?db=http://demo.noms.io&hash=abcd
is now:
  ?p=http://demo.noms.io::#abcd

and of course more complex paths work, like:
  ?p=http://demo.noms.io::#abcd.field[42].name

I've made 2 changes to the noms API for this:
1. Added support for an authorization token to DatabaseSpec.
2. Exposed getHashOfValue in noms.js.