Allow merge.ThreeWay() to operate across DBs
Much like Diff, we want to be able to merge values from different
databases. Unlike Diff, Merge needs to be able to resolve Refs --
which requires a ValueReadWriter that has access to the referenced
Value. So, merge.ThreeWay() now takes a ValueReadWriter for each Value
being considered for the merge -- the parent and both descendants.

Towards #148