Introduce noms-merge, a standalone noms tool for merging datasets
This is a first pass at exposing the new merge package to users.  The
tool is very basic, and currently only works on datasets in the same
database. It requires the 'parent' (i.e. a common ancestor of the two
datasets being merged) to be provided by a commandline option; a
follow-on patch will make the code discover this ancestor automatically.

Toward #2445