Confirm that overestimate of data written is consistent with leveldb stats
Did more investigation to determine whether the ~7-15% descrepency between
estimated bytes written and byte size of level-db directory is a concern.
One would expect the actual size on disk to be larger than estimate due to
leveldb overhead, but we're seeing the opposite. I've confirmed that
there no double counting of references. I've also confirmed that
the estimate consistent with stats reported using noms sync --lb-dump-stats.
This indicates that the issue is somewhere in the ldb driver and not a
noms bug.

fixes: #1998