Add noms commit --merge
Add optional merging functionality to noms commit.
noms commit --merge <database> <left-dataset-name> <right-dataset-name>

The command above will look in the given Database for the two named
Datasets and, if possible, merge their HeadValue()s and commit the
result back to <right-dataset-name>. All the existing options to
`noms commit` are supported, to allow users to add metadata to the
Commit object that winds up in the Database.

Users can optionally provide a third Dataset name to commit to instead.

Fixes #2535