Fix sync bug when all chunks are already present in the sink
In the edge case where all source chunks are already present in the
sink AND the sink dataset does not yet exist, the httpBatchStore code
was sending chunks in reverse order.  This patch ensures that,
regardless of how few chunks are sent, any operation that sends chunks
to the server also resets chunk write order.

Fixes #3101