Loosen checks in ValueStore.bufferChunks
Before ripping out all the hinting and associated proactive
validation checking, it was impossible to write Values through
the ValueStore API in any way other than bottom-up. That meant
that we could enforce the invariant that a chunk could not
appear in pendingParents unless it was also in pendingPuts.
Now, it's possible to construct legal sequences of calls to
ValueStore that result in this check being violated. Raf and
I don't think that violating this check can actually lead to
an invalid underlying database, as the lazy validation that
replaced the proactive validation should still catch any such
issues.

Fixes #3371